Implement basic Q-Learning
Create a new Q-Learner class encapsulating update() and optimal_action() methods. The class is flexible enough to allow different kinds of state and action space.

Also includes a rudimentary exploration decay factor.

The agent (or state space) still needs to be tweaked because it is finding high rewards in going around in circles --> Ex: +2 +2 +2 -0.5.