Make a working agent (weird one)
Experiment and find a weird working agent that
somehow uses the state "sensed" before taking
action and the state "sensed" before taking action
on the previous iteration with reward and action
for the next state to update Q-table. (Some
symmetry in the underlying model should be
responsible for this behavior.) Agent learns
to reliably (>98%) reach destination after 10 trials.

More sensible version of the agent and parameter
tuning is pending.