Make a working agent (proper one)
Implement such that the two states used to
update Q-table are the one corresponding
to before and after taking action in the same
iteration. This is more intuitive than the previous
version. Agent learns to reach destination
reliably after 10 trials.

Also, beautify the implementation a bit and make
the discount factor (gamma) and exploration
factor (epsilon) parameterizable.

Parameter tuning, and potentially update
of agent with optimal parameters pending.