Add agent (and simulator) with analysis tools
Update simulator to return some values that are
pertaining to each simulator run (100 trials of
the agent). The values are then returned to
agent.py, in which the main run() function is
called with several alpha, epsilon and gamma
parameters. In agent.py, the run(), main entry,
are all made to support both running the project
in "default" mode or to run it in analysis mode. Additionally make the simulations multi-process
as larger parameter spaces seem to take quite
a while.

Also the learning factor (alpha_prime) is added
to the agent.